persist: hook up user tables to persistence core behind flag
Revive the --persistent_tables flag which enables persistence of user
tables. This is nowhere near ready for even the most experimental of
users, so hide the flag. Really only doing this now to establish the
necessary plumbing and to make sure the interface of the persist crate
plays well with what coord/dataflow need.