persist: WIP strawman for new multi-writer persist interface
This is an attempt to redesign the persist API to balance all of the
following factors:
- Usage within STORAGE
- Operational simplicity
- Maintainability of implementation
- Performance
- Cost

Additionally, we're also taking this shift in product priorities (as
well as some assumptions that have changed since persist started) to lay
out a plan to incrementally ship this thing to production. Shipping
persist incrementally is something everyone wants, is much less risky,
and is an earlier timeline. This is something we've tried hard to do
historically, but have been unable to tease apart into a workable plan
(given the constraints of our initial goal of fast restarts for kafka
source persistence). Without over-promising, this interface rethink
seems at the moment like it may unlock incremental delivery.

That's the good news. The bonus good news is that persist's internal
implementation is pretty clean and contains lots of great building
blocks. Even if we entirely throw out the current API and switch to this
one, it should be a pretty fast process. This is in no way a rewrite.

Here's my rough short-to-medium-term plan:
- Figure out the balance of persist priorities (support platform vs the
  kafka source fast restart alpha) with product. Adjust the following
  steps accordingly. Timeline: ASAP
- Get persist and storage to agree on this API boundary between them,
  including performance and cost expectations. Then, get sign-off from
  Cloud on the operational implications. Merge it. Timeline: ASAP
- Write and merge a prototype-quality implementation of this API to
  prove the concept and unblock storage prototyping. Timeline: ASAP
- Write and merge a design doc for the production-quality implementation
  of this API and its incremental rollout. Timeline: ASAP
- Write a bare-bones production-quality implementation of this API. This
  will happen in parallel with the initial production-quality
  implementation of STORAGE. Adjust the API as necessary in light of the
  inevitable discovery of unforeseen issues+requirements as persist and
  STORAGE are implemented. Timeline: Platform launch.
- Continue to incrementally ship persist features as decided by product
  prioritization.