storage: add DataSourceOther::Shard
Splitting this out of #24138, which exposes the catalog shard via SQL,
because there are some open questions to resolve for the adapter pieces,
but the storage part is ready to go. Bonus, this neatly splits the
needed code reviewers.

The catalog is already a persist shard with the usual `SourceData, (),
Timestamp, Diff` type, maintained entirely outside storage. Model this
as a storage collection which happens to be backed by a persist shard.
We could copy this shard into one managed entirely by storage, but as a
performance and complexity optimization, we don't.